Продолжаю декомпозирование
